requirements.txt
Flask==3.1.2
Flask-Cors==3.1.1
SQLAlchemy==2.0.21
GeoAlchemy2==0.13.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
